split out calibration in pipette info
